epai/requirements.txt
streamlit>=1.32
langchain>=0.2
langchain-community>=0.2
langchain-ollama>=0.1
chromadb>=0.5
ollama>=0.2
pdfplumber>=0.11
pypdf>=4.0
pdf2image>=1.17
pytesseract>=0.3.10
Pillow>=10.0
openpyxl>=3.1
pandas>=2.0
reportlab>=4.0
sentence-transformers>=3.0
